rename: change icon name Share to Link as LinkIcon
