+++ src/components/icons/index.ts
@@ -3,7 +3,7 @@ export { default as Lead } from './Lead'
 export { default as Check } from './Check'
 export { default as Email } from './Email'
 export { default as Instagram } from './Instagram'
-export { default as Share } from './Share'
+export { default as Link } from './Link'
 export { default as Boy } from './Boy'
 export { default as Girl } from './Girl'
 export { default as Gdsc } from './Gdsc'
